+++ tidy3d/components/data/dataset.py
@@ -403,4 +403,3 @@ class TriangleMeshDataset(Dataset):
         description="Dataset containing the surface triangles and corresponding face indices "
         "for a surface mesh.",
     )
-