+++ tidy3d/components/simulation.py
@@ -281,11 +281,11 @@ class Simulation(Box):  # pylint:disable=too-many-public-methods
         """Error if there are plane wave sources incompatible with boundary conditions."""
         boundaries = val.to_list
         sources = values.get("sources")
+        if not sources:
+            return val
         size = values.get("size")
         sim_medium = values.get("medium")
         structures = values.get("structures")
-        if not structures:
-            return val
         for source in sources:
             if not isinstance(source, PlaneWave):
                 continue
@@ -323,12 +323,12 @@ class Simulation(Box):  # pylint:disable=too-many-public-methods
         """Error if the boundary conditions are compatible with TFSF sources, if any."""
         boundaries = val.to_list
         sources = values.get("sources")
+        if not sources:
+            return val
         size = values.get("size")
         center = values.get("center")
         sim_medium = values.get("medium")
         structures = values.get("structures")
-        if not structures:
-            return val
         sim_bounds = [
             [c - s / 2.0 for c, s in zip(center, size)],
             [c + s / 2.0 for c, s in zip(center, size)],
@@ -715,8 +715,6 @@ class Simulation(Box):  # pylint:disable=too-many-public-methods
         """If any :class:`.DiffractionMonitor` exists, ensure is does not lie in a lossy medium."""
         monitors = val
         structures = values.get("structures")
-        if not structures:
-            return val
         medium = values.get("medium")
         for monitor in monitors:
             if isinstance(monitor, DiffractionMonitor):